Break up the `game` route` into multiple functions
Handle `game_board` and `game_status` as independent concerns.
